requirements.txt
MetaTrader5
pandas
pandas_ta
requests
python-dotenv
